% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_comic_id}
\alias{get_comic_id}
\title{Get comic ID from tag}
\usage{
get_comic_id(tag)
}
\arguments{
\item{tag}{Character string specifying the keyword tag (e.g.,
\code{"birthday"}).}
}
\description{
Use a JSON file to look up a comic ID associated with a given tag.
}
\examples{
get_comic_id("birthday")
get_comic_id("edvard")  # Edvard Munch's "The Scream"
}
