% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_friends_comic_ids}
\alias{get_friends_comic_ids}
\title{Get comic IDs from tag}
\usage{
get_friends_comic_ids(tag)
}
\arguments{
\item{tag}{Character string specifying the keyword tag (e.g.,
\code{"birthday"}).}
}
\description{
Use a JSON file to look up all comic IDs associated with a given tag.
}
\examples{
get_friends_comic_ids("birthday")
get_friends_comic_ids("edvard munch")  # Edvard Munch's "The Scream"
}
