% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_id}
\alias{get_id}
\title{Get a User Id}
\usage{
get_id(user_email)
}
\arguments{
\item{user_email}{Character string specifying the email associated with the Bitmoji account
for which you'd like to get/plot comics (e.g., \code{username@gmail.com}).}
}
\description{
Get an \code{id} for consumption by \code{\link{plot_comic}} and \code{\link{get_comic}}
}
\examples{
\dontrun{
id <- get_id("username@gmail.com")
plot_comic(paste0(id,"-v1"), tag = "edvard")
}
}
