% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comic.R
\name{get_comic}
\alias{get_comic}
\title{Get Bitmoji comic}
\usage{
get_comic(id, tag)
}
\arguments{
\item{id}{Character string specifying the users unique ID. Can also be a
vector of length 2.}

\item{tag}{Character string specifying the keyword tag (e.g.,
\code{"birthday"}).}
}
\description{
Get a Bitmoji comic.
}
\examples{
# ID obtained using get_id()
my_id <- "1551b314-5e8a-4477-aca2-088c05963111-v1"
get_comic(my_id, tag = "edvard")
get_comic(c(my_id, my_id), tag = "edvard")
}
