% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get user id}
\usage{
get_id(user_email)
}
\arguments{
\item{user_email}{Character string specifying the email associated with the
Bitmoji account for which you'd like to get/plot comics (e.g.,
\code{username@gmail.com}).}
}
\description{
Get unique user id for use with \code{\link{plot_comic}} and
\code{\link{get_comic}}.
}
\examples{
\dontrun{
my_id <- get_id("username@gmail.com")
plot_comic(my_id, tag = "edvard")
}
}
