% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comic.R
\name{plot_comic}
\alias{plot_comic}
\title{Plot Bitmoji comic}
\usage{
plot_comic(id, tag = NULL, comic_id = NULL)
}
\arguments{
\item{id}{Character string specifying the users unique ID. Can also be a
vector of length 2.}

\item{tag}{Character string specifying the keyword tag (e.g.,
\code{"birthday"}).}

\item{comic_id}{Character string specifying the commic id (e.g., \code{"10228108"})}
}
\description{
Plot a Bitmoji comic.
}
\examples{
# ID obtained using get_id()
my_id <- "1551b314-5e8a-4477-aca2-088c05963111-v1"
plot_comic(my_id, tag = "edvard")

# Plot with friends
plot_comic(c(my_id, my_id), tag = "sloth")
}
